content/_code-samples/build-a-wallet/py/requirements.txt
xrpl-py==1.1.1
wxPython=4.1.1
